refactor: Extract directive parsing and merge logic from renderers.js into utils/directives.js
